Fix tests failing due to new release of sklearn (0.21.0) - Enforce int type labels from stream generators - Reset sklearn version in requirements to >=0.20 - Closes #114
Former-commit-id: 173f2688750e194413e99bf4a13ed7f9a0adb712

+++ src/skmultiflow/data/hyper_plane_generator.py
@@ -236,7 +236,7 @@ class HyperplaneGenerator(Stream):
         self._generate_drift()
 
         self.current_sample_x = data[:, :self.n_features]
-        self.current_sample_y = data[:, self.n_features:].flatten()
+        self.current_sample_y = data[:, self.n_features:].flatten().astype(int)
 
         return self.current_sample_x, self.current_sample_y
