+++ src/skmultiflow/data/mixed_generator.py
@@ -206,7 +206,7 @@ class MIXEDGenerator(Stream):
             data[j, 4] = group
 
         self.current_sample_x = data[:, :self.n_features]
-        self.current_sample_y = data[:, self.n_features:].flatten()
+        self.current_sample_y = data[:, self.n_features:].flatten().astype(int)
 
         return self.current_sample_x, self.current_sample_y
