+++ src/skmultiflow/data/sea_generator.py
@@ -261,7 +261,7 @@ class SEAGenerator(Stream):
             data[j, 3] = group
 
         self.current_sample_x = data[:, :self.n_features]
-        self.current_sample_y = data[:, self.n_features:].flatten()
+        self.current_sample_y = data[:, self.n_features:].flatten().astype(np.int64)
 
         return self.current_sample_x, self.current_sample_y
